#4 - refactoring role validation
